fix payment status

+++ database/migrations/2025_05_09_110200_create_orders_table.php
@@ -14,10 +14,10 @@ return new class extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table->bigInteger('user_id')->unsigned();
-            $table->decimal('subtotal');
-            $table->decimal('discount')->default(0);
-            $table->decimal('tax');
-            $table->decimal('total');
+            $table->int('subtotal');
+            $table->int('discount')->default(0);
+            $table->int('tax');
+            $table->int('total');
             $table->string('name');
             $table->string('phone');
             $table->string('locality');
